Split tabor tests based on their backend(simulator or dummy module); Clarify some driver methods
